fix: make create as draft, only create as draft
The option was not doing what it said.

They is no point to change the draft status once the PR has been
submitted.

This was changing the draft status set in GitHub UI each time we run
Mergify cli

Now we only set the draft status when a PR is created as the doc string
say.