fix: don't recreate merged PR
If skip-rebase is used, and first PR of the stack is merged, we recreate
it because the pull request has vanish remotely.

This doesn't make sense, this change instead mark it locally as merged.

The merged status will displayed in cli output until the next rebase.

This avoids creating useless PR.

Change-Id: Idfe99f64e6f9bdb788c8120173479f77fffbb828