fix: change the way we reexec mergify to enable PYTHONUTF8 (#777)
poetry/setuptools creates a windows binary on windows.

The current reexec code does not work with this binary as we expect the
sys.argv[0] to be python script.

This changes the way we enable UTF8 mode to reexec the exact binary that
launch the initial process.

This should works on all platforms whatever the console script file is a
python file or a binary.