fix: modify commit-msg hook to work with verbose commits (#782)
Small suggestion for the commit-msg hook to strip everything below the scissor line (added by [`git commit -v`](https://git-scm.com/docs/git-commit#Documentation/git-commit.txt---verbose)) before stripping all of the comments. If this doesn't get stripped the commit message will never be detected as empty because of the diff below the scissor line (which are not comments).

This change is merely a suggestion, feel free to take it and modify it as you see fit. Preferably it would also check the `--cleanup`/`git config commit.cleanup` option and only strip as much as that specifies but that's maybe something for in the future.

Tested `git commit` with and without `-v`, and with and without a commit message.