fix: make a json dict for the GitHub outputs
GitHub supports dynamic outputs only if the action use "using: node".
This does not work for "using: composite" or "docker" action.

The workaround is to dump the json into the output.

Then the GitHub workflow will be able to use fromJSON to read it:

```
if: fromJSON(steps.action.outputs.scopes).everything
```

Change-Id: I294c5f828801ad06a6ed613a127ed1b9315e40aa