fix: make a json dict for the GitHub output (#829)
GitHub supports dynamic outputs only if the action use "using: node".
This does not work for "using: composite" or "docker" action.

https://github.com/orgs/community/discussions/10529

The workaround is to dump the json into the output.

Then the GitHub workflow will be able to use fromJSON to read it:

```
if: fromJSON(steps.action.outputs.scopes).everything
```