feat: more accurate head detection (#843)
We currently uses HEAD to the tip of the PR branch.
When it's a PR, the HEAD is a merge_commit, this merge commits may
contains file changes from main, but not from the PR.

This change adjusts the code to use head.sha instead.

This change the changed_files algo more accurate.